+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.53.1] - 2024-10-08
+
+### Fixed
+
+- Keyboard control in NVDA form mode for Legend checkbox items.
+- Each checkbox in the Legend rendered as 3 elements for NVDA.
+
 ## [3.53.0] - 2024-10-08
 
 ### Added